Ported Logins model over to kohana ORM; Reworked AuthProfiles cookie handling to detect disabled logins and keep cached data current with DB; Switched to objects from arrays for all DB results; Updated MySQL schema, using foreign key constraints and cascading deletes;

+++ modules/auth_profiles/models/login.php
@@ -1,18 +1,24 @@
 <?php
 /**
- * Logins model
+ * Login model
  *
- * @package    Memex
+ * @package    auth_profiles
  * @subpackage models
  * @author     l.m.orchard <l.m.orchard@pobox.com>
  */
-class Logins_Model extends Model
+class Login_Model extends ORM
 {
-    protected $_table_name = 'logins';
+    // {{{ Model attributes
+    
+    protected $has_and_belongs_to_many = array('profiles');
+    protected $has_many = array('profile_attributes');
+
     protected $_table_name_password_reset_token =
         'login_password_reset_tokens';
     protected $_table_name_email_verification_token =
         'login_email_verification_tokens';
+
+    // }}}
 
     /**
      * One-way encrypt a plaintext password, both for storage and comparison 
@@ -27,193 +33,80 @@ class Logins_Model extends Model
     }
 
     /**
-     * Create a new login
+	 * Returns the unique key for a specific value. This method is expected
+	 * to be overloaded in models if the model has other unique columns.
+	 *
+     * If the key used in a find is a non-numeric string, search 'login_name' column.
      *
-     * @param array Login data
-     * @return string New login ID
+	 * @param   mixed   unique value
+	 * @return  string
      */
-    public function create($data)
+    public function unique_key($id)
     {
-        if (empty($data['login_name']))
-            throw new Exception('login_name required');
-        if (empty($data['email']))
-            throw new Exception('email required');
-        if (empty($data['password']))
-            throw new Exception('password required');
-        if ($this->find_by_login_name($data['login_name']))
-            throw new Exception('duplicate login name');
-
-        $data = array(
-            'login_name' => $data['login_name'],
-            'email'      => $data['email'],
-            'password'   => $this->encrypt_password($data['password']),
-            'created'    => gmdate('Y-m-d H:i:s', time())
-        );
-        $data['id'] = $this->db
-            ->insert($this->_table_name, $data)
-            ->insert_id();
-
-        return $data;
+        if (!empty($id) && is_string($id) && !ctype_digit($id)) {
+            return 'login_name';
+        }
+        return parent::unique_key($id);
     }
 
     /**
-     * Delete a login.  
+     * Before saving, update created/modified timestamps and generate a UUID if 
+     * necessary.
      *
-     * Note that this does not cascadingly delete profiles or anything else, 
-     * since profiles are the primary resource here and multiple logins may be 
-     * attached to a single profile.
-     *
-     * @param string Login ID
+	 * @chainable
+	 * @return  ORM
      */
-    public function delete($id) {
-        $this->db->delete($this->_table_name, array('id'=>$id));
+    public function save()
+    {
+        if (isset($this->table_columns['password']) && !empty($this->password)) {
+            $this->password = $this->encrypt_password($this->password);
+        }
+        return parent::save();
     }
+
 
     /**
      * Create a new login and associated profile.
      */
-    public function register_with_profile($data)
+    public function register_with_profile($data, $force_email_verified=false)
     {
-        $new_login = $this->create($data);
-        try {
-            $profiles_model = new Profiles_Model();
-            $new_profile = $profiles_model->create($data);
-            $this->add_profile_to_login($new_login['id'], $new_profile['id']);
-        } catch (Exception $e) {
-            // If profile creation failed, delete the login.
-            // TODO: Transaction here?
-            $this->delete($new_login['id']);
-            throw $e;
-        }
-        return $new_login;
-    }
-
-    /**
-     * Link an profile with this login
-     */
-    public function add_profile_to_login($login_id, $profile_id) 
-    {
-        return $this->db->insert(
-            'logins_profiles', array(
-                'login_id'   => $login_id, 
-                'profile_id' => $profile_id
-            )
+        $profile_data = array(
+            'login_name' => $data['login_name'],
+            'email'      => ($force_email_verified) ? $data['email'] : '',
+            'password'   => $data['password'],
+            'created'    => gmdate('c', time())
         );
+
+        $new_login = ORM::factory('login')->set($profile_data)->save();
+
+        $profile_data['id'] = $new_login->id;
+
+        if (!$force_email_verified) {
+            $profile_data['new_email'] = $data['email'];
+            $profile_data['email_verification_token'] = 
+                $this->set_email_verification_token(
+                    $new_login->id, $data['email']
+                );
+        }
+
+        $new_profile = ORM::factory('profile')->set($data)->save();
+
+        $new_login->add($new_profile);
+        $new_login->save();
+
+        return arr::to_object($profile_data);
     }
 
     /**
-     * Look up by login name
-     *
-     * @param string Screen name
+     * Find the default profile for this login, usually the first registered.
      */
-    public function find_by_login_name($login_name)
+    public function find_default_profile_for_login()
     {
-        $row = $this->db->select()->from($this->_table_name)
-            ->where('login_name', $login_name)
-            ->get()->current();
-        if (!$row) return null;
-        return $row;
+        if (!$this->loaded) return null;
+        $profiles = $this->profiles;
+        return $profiles[0];
     }
 
-    /**
-     * Look up by email
-     *
-     * @param string email
-     */
-    public function find_by_email($email)
-    {
-        $row = $this->db->select()->from($this->_table_name)
-            ->where('email', $email)
-            ->get()->current();
-        if (!$row) return null;
-        return $row;
-    }
-
-    /**
-     * Look up by reset token
-     *
-     * @param string Screen name
-     * @
-     */
-    public function find_by_password_reset_token($token)
-    {
-        $row = $this->db->select($this->_table_name . '.*')
-            ->from($this->_table_name)
-            ->join(
-                $this->_table_name_password_reset_token,
-                $this->_table_name_password_reset_token . '.login_id',
-                $this->_table_name . '.id'
-            )
-            ->where(
-                $this->_table_name_password_reset_token . '.token', 
-                $token
-            )
-            ->get()->current();
-        if (!$row) return null;
-        return $row;
-    }
-
-    /**
-     * Look up by reset token
-     *
-     * @param string Screen name
-     * @
-     */
-    public function find_by_email_verification_token($token)
-    {
-        $row = $this->db->select(
-                $this->_table_name . '.*',
-                $this->_table_name_email_verification_token . '.value AS new_email'
-            )
-            ->from($this->_table_name)
-            ->join(
-                $this->_table_name_email_verification_token,
-                $this->_table_name_email_verification_token . '.login_id',
-                $this->_table_name . '.id'
-            )
-            ->where(
-                $this->_table_name_email_verification_token . '.token', 
-                $token
-            )
-            ->get()->current();
-        if (!$row) return null;
-        return $row;
-    }
-
-    /**
-     * Fetch the default profile for a login.
-     */
-    public function find_default_profile_for_login($login_id)
-    {
-        $profiles = $this->find_profiles_for_login($login_id);
-        return (!$profiles) ? null : $profiles[0];
-    }
-
-    /**
-     * Get all profiles for a login
-     */
-    public function find_profiles_for_login($login_id)
-    {
-        $login_row = $this->db->select()
-            ->from($this->_table_name)
-            ->where('id', $login_id)
-            ->get()->current();
-
-        if (null == $login_row) return null;
-
-        $profile_rows = $this->db
-            ->select('profiles.*')
-            ->from('profiles')
-            ->join('logins_profiles', 'logins_profiles.profile_id=profiles.id')
-            ->where('logins_profiles.login_id', $login_row['id'])
-            ->get()->result_array();
-
-        $profiles = array();
-        foreach ($profile_rows as $row)
-            $profiles[] = $row;
-
-        return $profiles;
-    }
 
     /**
      * Set the password reset token for a given login and return the value 
@@ -222,18 +115,20 @@ class Logins_Model extends Model
      * @param  string login ID
      * @return string password reset string
      */
-    public function set_password_reset_token($login_id)
+    public function set_password_reset_token()
     {
+        if (!$this->loaded) return;
+
         $token = md5(uniqid(mt_rand(), true));
 
         $this->db->delete(
             $this->_table_name_password_reset_token,
-            array( 'login_id' => $login_id )
+            array( 'login_id' => $this->id )
         );
         $rv = $this->db->insert(
             $this->_table_name_password_reset_token,
             array(
-                'login_id' => $login_id,
+                'login_id' => $this->id,
                 'token'    => $token
             )
         );
@@ -249,18 +144,20 @@ class Logins_Model extends Model
      * @param  string  new password value
      * @return boolean whether or not a password was changed
      */
-    public function change_password($login_id, $new_password)
+    public function change_password($new_password)
     {
+        if (!$this->loaded) return;
+
         $crypt_password = $this->encrypt_password($new_password);
 
         $this->db->delete(
             $this->_table_name_password_reset_token,
-            array( 'login_id' => $login_id )
+            array( 'login_id' => $this->id )
         );
         $rows = $this->db->update(
             'logins', 
             array('password'=>$crypt_password), 
-            array('id'=>$login_id)
+            array('id'=>$this->id)
         );
 
         return !empty($rows);
@@ -273,18 +170,20 @@ class Logins_Model extends Model
      * @param  string login ID
      * @return string password reset string
      */
-    public function set_email_verification_token($login_id, $new_email)
+    public function set_email_verification_token($new_email)
     {
+        if (!$this->loaded) return;
+
         $token = md5(uniqid(mt_rand(), true));
 
         $this->db->delete(
             $this->_table_name_email_verification_token,
-            array( 'login_id' => $login_id )
+            array( 'login_id' => $this->id )
         );
         $rv = $this->db->insert(
             $this->_table_name_email_verification_token,
             array(
-                'login_id' => $login_id,
+                'login_id' => $this->id,
                 'token'    => $token,
                 'value'    => $new_email
             )
@@ -301,20 +200,67 @@ class Logins_Model extends Model
      * @param  string  new email value
      * @return boolean whether or not a email was changed
      */
-    public function change_email($login_id, $new_email)
+    public function change_email($new_email)
     {
+        if (!$this->loaded) return;
         $this->db->delete(
             $this->_table_name_email_verification_token,
-            array( 'login_id' => $login_id )
+            array( 'login_id' => $this->id )
         );
         $rows = $this->db->update(
             'logins', 
             array('email'=>$new_email), 
-            array('id'=>$login_id)
+            array('id'=>$this->id)
         );
-
         return !empty($rows);
     }
+
+
+    /**
+     * Find by password reset token
+     *
+     * @param  string token value
+     * @return Login_Model
+     */
+    public function find_by_password_reset_token($token)
+    {
+        return ORM::factory('login')
+            ->join(
+                $this->_table_name_password_reset_token,
+                "{$this->_table_name_password_reset_token}.login_id",
+                "{$this->table_name}.id"
+            )
+            ->where(
+                "{$this->_table_name_password_reset_token}.token",
+                $token
+            )
+            ->find();
+    }
+
+    /**
+     * Find by email verification token
+     *
+     * @param  string token value
+     * @return Login_Model
+     */
+    public function find_by_email_verification_token($token)
+    {
+        $row = $this->db
+            ->select('value, login_id')
+            ->from($this->_table_name_email_verification_token)
+            ->where('token', $token)
+            ->get()->current();
+
+        if (!$row) {
+            return array(null, null);
+        } else {
+            return array(
+                ORM::factory('login', $row->login_id),
+                $row->value
+            );
+        }
+    }
+
 
     /**
      * Replace incoming data with registration validator and return whether 
@@ -325,19 +271,23 @@ class Logins_Model extends Model
      */
     public function validate_registration(&$data)
     {
-        $profiles_model = new Profiles_Model();
+        $profile_model = new Profile_Model();
 
         $data = Validation::factory($data)
             ->pre_filter('trim')
             ->add_rules('login_name',       
                 'required', 'length[3,64]', 'valid::alpha_dash', 
                 array($this, 'is_login_name_available'))
-            ->add_rules('email', 'required', 'valid::email')
+            ->add_rules('email', 
+                'required', 'length[3,255]', 'valid::email',
+                array($this, 'is_email_available'))
+            ->add_rules('email_confirm', 
+                'required', 'valid::email', 'matches[email]')
             ->add_rules('password', 'required')
             ->add_rules('password_confirm', 'required', 'matches[password]')
             ->add_rules('screen_name',      
                 'required', 'length[3,64]', 'valid::alpha_dash', 
-                array($profiles_model, 'isScreenNameAvailable'))
+                array($profile_model, 'is_screen_name_available'))
             ->add_rules('full_name', 'required', 'valid::standard_text')
             ->add_rules('captcha', 'required', 'Captcha::valid')
             ;
@@ -357,7 +307,8 @@ class Logins_Model extends Model
     {
         $data = Validation::factory($data)
             ->pre_filter('trim')
-            ->add_rules('login_name', 'required', 'length[3,64]', 'valid::alpha_dash')
+            ->add_rules('login_name', 'required', 'length[3,64]', 
+                'valid::alpha_dash', array($this, 'is_login_name_registered'))
             ->add_rules('password', 'required')
             ->add_callbacks('password', array($this, 'is_password_correct'))
             ;
@@ -376,7 +327,8 @@ class Logins_Model extends Model
         $data = Validation::factory($data)
             ->pre_filter('trim')
             ->add_rules('new_email', 
-                'required', 'valid::email')
+                'required', 'length[3,255]', 'valid::email',
+                array($this, 'is_email_available'))
             ;
         return $data->validate();
     }
@@ -440,14 +392,17 @@ class Logins_Model extends Model
         return $data->validate();
     }
 
+
     /**
      * Check to see whether a login name is available, for use in form 
      * validator.
      */
     public function is_login_name_available($name)
     {
-        $login = $this->find_by_login_name($name);
-        return empty($login);
+        $count = $this->db
+            ->where('login_name', $name)
+            ->count_records($this->table_name);
+        return (0==$count);
     }
 
     /**
@@ -463,9 +418,19 @@ class Logins_Model extends Model
      * Check to see whether a given email address has been registered to a 
      * login, for use in form validation.
      */
+    public function is_email_available($email) {
+        $count = $this->db
+            ->where('email', $email)
+            ->count_records($this->table_name);
+        return (0==$count);
+    }
+
+    /**
+     * Check to see whether a given email address has been registered to a 
+     * login, for use in form validation.
+     */
     public function is_email_registered($email) {
-        $login = $this->find_by_email($email);
-        return !(empty($login));
+        return !($this->is_email_available($email));
     }
 
     /**
@@ -476,9 +441,11 @@ class Logins_Model extends Model
     {
         $login_name = (isset($valid['login_name'])) ?
             $valid['login_name'] : AuthProfiles::get_login('login_name');
-        $login = $this->find_by_login_name($login_name);
-        if ($this->encrypt_password($valid[$field]) != $login['password'])
-            $valid->add_error($field, 'invalid');
+        $count = $this->db
+            ->where('login_name', $login_name)
+            ->where('password', $this->encrypt_password($valid[$field]))
+            ->count_records($this->table_name);
+        return !(0===$count);
     }
 
     /**
@@ -487,11 +454,12 @@ class Logins_Model extends Model
      * @param  string  password reset token
      * @return boolean 
      */
-    public function is_password_reset_token_valid($reset_token)
+    public function is_password_reset_token_valid($token)
     {
-        // TODO: Do a count() query or something simpler.
-        $login = $this->find_by_password_reset_token($reset_token);
-        return !empty($login);
+        $count = $this->db
+            ->where('token', $token)
+            ->count_records($this->_table_name_password_reset_token);
+        return !(0==$count);
     }
 
     /**
@@ -520,17 +488,6 @@ class Logins_Model extends Model
                 $valid->add_error($field, 'default');
             }
         }
-    }
-
-    /**
-     * Delete all users from the system.  Useful for tests, but dangerous 
-     * otherwise.
-     */
-    public function delete_all()
-    {
-        if (!Kohana::config('model.enable_delete_all'))
-            throw new Exception('Mass deletion not enabled');
-        $this->db->query('DELETE FROM ' . $this->_table_name);
     }
 
 }
